+++ module-api/src/main/java/com/numo/api/comm/page/PageResponse.java
@@ -1,0 +1,18 @@
+package com.numo.api.comm.page;
+
+import org.springframework.data.domain.Slice;
+
+import java.util.List;
+
+public record PageResponse<T>(
+        PageDto page,
+        List<T> data
+){
+    public PageResponse(Slice<T> data) {
+        this(data, null);
+    }
+
+    public PageResponse(Slice<T> data, Long lastId) {
+        this(new PageDto(data.getNumber(), data.hasNext(), lastId), data.getContent());
+    }
+}
