--- module-api/src/main/java/com/numo/api/dto/post/ReadPostListResponseDto.java
@@ -1,11 +1,0 @@
-package com.numo.api.dto.post;
-
-import com.numo.api.dto.page.PageDto;
-
-import java.util.List;
-
-public record ReadPostListResponseDto(
-        PageDto page,
-        List<PostListResponseDto> posts
-) {
-}
